requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
argilla==1.3.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
beautifulsoup4==4.11.2
black==23.1.0
blinker==1.5
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.3.0
entrypoints==0.4
et-xmlfile==1.1.0
faiss-cpu==1.7.3
fastapi==0.93.0
filelock==3.9.0
frozenlist==1.3.3
geojson==2.5.0
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.1
idna==3.4
importlib-metadata==6.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.125
libmagic==1.0
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.23.5
openai==0.27.1
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
pinecone-client==2.2.1
platformdirs==3.1.0
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.6
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyowm==3.3.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
rich==13.3.2
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.46
starlette==0.25.0
streamlit-chat==0.0.2.2
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
unstructured==0.5.2
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
websockets==10.4
wrapt==1.14.1
XlsxWriter==3.0.8
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
